refactor: remove unnecessary `pub mod` declarations
Modules were already exposed via `pub use`, making `pub mod` redundant. Changed
all instances to private `mod`. No functional changes.
