+++ src/actions.rs
@@ -1,6 +1,6 @@
-pub mod create_instance_action;
-pub mod start_instance_action;
-pub mod stop_instance_action;
+mod create_instance_action;
+mod start_instance_action;
+mod stop_instance_action;
 
 pub use create_instance_action::CreateInstanceAction;
 pub use start_instance_action::StartInstanceAction;
