+++ src/commands.rs
@@ -1,27 +1,27 @@
-pub mod clone_command;
-pub mod command_dispatcher;
-pub mod completions_command;
-pub mod console_command;
-pub mod create_command;
-pub mod delete_command;
-pub mod exec_command;
-pub mod image;
-pub mod list_image_command;
-pub mod list_instance_command;
-pub mod list_port_command;
-pub mod modify_command;
-pub mod prune_command;
-pub mod rename_command;
-pub mod restart_command;
-pub mod run_command;
-pub mod scp_command;
-pub mod show_command;
-pub mod show_image_command;
-pub mod show_instance_command;
-pub mod ssh_command;
-pub mod start_command;
-pub mod stop_command;
-pub mod verbosity;
+mod clone_command;
+mod command_dispatcher;
+mod completions_command;
+mod console_command;
+mod create_command;
+mod delete_command;
+mod exec_command;
+mod image;
+mod list_image_command;
+mod list_instance_command;
+mod list_port_command;
+mod modify_command;
+mod prune_command;
+mod rename_command;
+mod restart_command;
+mod run_command;
+mod scp_command;
+mod show_command;
+mod show_image_command;
+mod show_instance_command;
+mod ssh_command;
+mod start_command;
+mod stop_command;
+mod verbosity;
 
 pub use clone_command::*;
 pub use command_dispatcher::*;
