+++ src/commands/list_instance_command.rs
@@ -82,11 +82,11 @@ impl Command for ListInstanceCommand {
 mod tests {
     use super::*;
     use crate::arch::Arch;
-    use crate::image::image_store_mock::tests::ImageStoreMock;
+    use crate::image::ImageStoreMock;
     use crate::instance::Instance;
-    use crate::instance::instance_store_mock::tests::InstanceStoreMock;
+    use crate::instance::InstanceStoreMock;
     use crate::model::DataSize;
-    use crate::view::console_mock::tests::ConsoleMock;
+    use crate::view::ConsoleMock;
 
     #[test]
     fn test_list_instance_command() {
