+++ src/env.rs
@@ -1,5 +1,5 @@
-pub mod environment;
-pub mod environment_factory;
+mod environment;
+mod environment_factory;
 
 pub use environment::Environment;
 pub use environment_factory::EnvironmentFactory;
