+++ src/model.rs
@@ -1,5 +1,5 @@
-pub mod data_size;
-pub mod instance_image_name;
+mod data_size;
+mod instance_image_name;
 
 pub use data_size::*;
 pub use instance_image_name::*;
